AI/codex_ai_pipeline/requirements.txt
langchain==0.1.0
langsmith>=0.0.77
chromadb==0.4.18
sentence-transformers==2.2.2
openai==1.3.0
anthropic==0.7.0
pinecone-client==2.2.4
whisper==1.1.10
elevenlabs==0.2.26
slack-sdk==3.23.0
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
prometheus-client==0.19.0
wandb==0.16.0
pydantic==2.5.0
python-dotenv==1.0.0
